services/indexer/requirements.txt
kafka-python==2.0.2
qdrant-client==1.7.0
psycopg2-binary==2.9.9
